MAT-8747 refactor code
