+++ src/main/java/cms/gov/madie/measure/locks/TestCaseLock.java
@@ -1,4 +1,4 @@
-package cms.gov.madie.measure.dto;
+package cms.gov.madie.measure.locks;
 
 import java.time.Instant;
